requirements.txt
weaviate-client
python
loguru
numpy
python-dotenv